Разбиение ScanView на четыре sub-controls
- Выделены ScanToolbar (диск, кнопки сканирования), ScanProgressPanel (живой прогресс), ScanSummaryStrip (итоги, слайдер), ScanStructurePanel (дерево, treemap)
- Локальные ресурсы (конвертеры, шаблоны, обработчики дерева) перенесены в соответствующие контролы
- ScanView.xaml переделана в тонкую оболочку (сетка, логика инспектора); новые контролы наследуют DataContext от ScanViewModel
- Плашка отмены и логика сплиттера оставлены в оболочке
