+++ SpaceSnoop.Wpf/Views/Controls/ScanProgressPanel.xaml
@@ -1,0 +1,134 @@
+﻿<UserControl x:Class="SpaceSnoop.Wpf.Views.Controls.ScanProgressPanel"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:vm="clr-namespace:SpaceSnoop.Wpf.ViewModels"
+             xmlns:behaviors="clr-namespace:SpaceSnoop.Wpf.Views.Behaviors"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d"
+             d:DataContext="{d:DesignInstance vm:ScanViewModel}"
+             d:DesignWidth="980">
+    <Border Style="{StaticResource Card.Inset}"
+            Padding="15,12"
+            Visibility="{Binding IsScanning, Converter={StaticResource BoolToVisibility}, FallbackValue=Collapsed}">
+        <StackPanel>
+            <Grid>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="Auto" />
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+
+                <StackPanel Grid.Column="0"
+                            Orientation="Horizontal"
+                            VerticalAlignment="Center">
+                    <Ellipse x:Name="ScanPulse"
+                             Width="8"
+                             Height="8"
+                             VerticalAlignment="Center"
+                             Margin="0,0,9,0"
+                             Fill="{DynamicResource Accent.Primary}">
+                        <Ellipse.Triggers>
+                            <EventTrigger RoutedEvent="Ellipse.Loaded">
+                                <BeginStoryboard>
+                                    <Storyboard>
+                                        <DoubleAnimation Storyboard.TargetProperty="Opacity"
+                                                         From="1"
+                                                         To="0.25"
+                                                         Duration="0:0:0.75"
+                                                         AutoReverse="True"
+                                                         RepeatBehavior="Forever" />
+                                    </Storyboard>
+                                </BeginStoryboard>
+                            </EventTrigger>
+                        </Ellipse.Triggers>
+                    </Ellipse>
+                    <TextBlock Text="СКАНИРОВАНИЕ"
+                               Style="{StaticResource Caption}"
+                               Foreground="{DynamicResource Accent.Primary}" />
+                </StackPanel>
+
+                <TextBlock Grid.Column="1"
+                           behaviors:PathTrimmer.FullPath="{Binding ScanCurrentPath}"
+                           VerticalAlignment="Center"
+                           Margin="14,0,14,0"
+                           TextWrapping="NoWrap"
+                           FontFamily="{DynamicResource Font.Mono}"
+                           FontSize="{DynamicResource Font.Size.S}"
+                           Foreground="{DynamicResource Fg.Secondary}"
+                           ToolTip="{Binding ScanCurrentPath}" />
+
+                <TextBlock Grid.Column="2"
+                           Text="{Binding ScanElapsedText}"
+                           Style="{StaticResource StatValue}"
+                           Foreground="{DynamicResource Accent.Primary}" />
+            </Grid>
+
+            <Grid Margin="0,11,0,0">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <ProgressBar Grid.Column="0"
+                             Height="6"
+                             VerticalAlignment="Center"
+                             Minimum="0"
+                             Maximum="1"
+                             Value="{Binding ProgressValue, Mode=OneWay}"
+                             IsIndeterminate="{Binding IsIndeterminate}" />
+                <TextBlock Grid.Column="1"
+                           Text="{Binding ScanPercentText}"
+                           MinWidth="40"
+                           Margin="10,0,0,0"
+                           TextAlignment="Right"
+                           VerticalAlignment="Center"
+                           FontFamily="{DynamicResource Font.Mono}"
+                           FontSize="{DynamicResource Font.Size.S}"
+                           FontWeight="Medium"
+                           Foreground="{DynamicResource Accent.Primary}"
+                           Visibility="{Binding ScanHasDeterminateProgress, Converter={StaticResource BoolToVisibility}, FallbackValue=Collapsed}" />
+            </Grid>
+
+            <StackPanel Orientation="Horizontal"
+                        Margin="0,11,0,0"
+                        VerticalAlignment="Center">
+                <TextBlock Text="КАТАЛОГОВ"
+                           Style="{StaticResource Caption}"
+                           Margin="0,0,7,0" />
+                <TextBlock Text="{Binding ScanDirCountText}"
+                           Style="{StaticResource StatValue}"
+                           Margin="0,0,18,0" />
+
+                <TextBlock Text="ФАЙЛОВ"
+                           Style="{StaticResource Caption}"
+                           Margin="0,0,7,0" />
+                <TextBlock Text="{Binding ScanFileCountText}"
+                           Style="{StaticResource StatValue}"
+                           Margin="0,0,18,0" />
+
+                <TextBlock Text="ОБЪЁМ"
+                           Style="{StaticResource Caption}"
+                           Margin="0,0,7,0" />
+                <TextBlock Text="{Binding ScanBytesText}"
+                           Style="{StaticResource StatValue}"
+                           Foreground="{DynamicResource Accent.Primary}"
+                           Margin="0,0,18,0" />
+
+                <TextBlock Text="СКОРОСТЬ"
+                           Style="{StaticResource Caption}"
+                           Margin="0,0,7,0" />
+                <TextBlock Text="{Binding ScanThroughputText}"
+                           Style="{StaticResource StatValue}"
+                           Margin="0,0,18,0" />
+
+                <TextBlock Text="ВЕТКИ"
+                           Style="{StaticResource Caption}"
+                           Margin="0,0,7,0"
+                           Visibility="{Binding ScanHasBranches, Converter={StaticResource BoolToVisibility}, FallbackValue=Collapsed}" />
+                <TextBlock Text="{Binding ScanTopLevelText}"
+                           Style="{StaticResource StatValue}"
+                           Visibility="{Binding ScanHasBranches, Converter={StaticResource BoolToVisibility}, FallbackValue=Collapsed}" />
+            </StackPanel>
+        </StackPanel>
+    </Border>
+</UserControl>
