+++ SpaceSnoop.Wpf/Views/Controls/ScanProgressPanel.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace SpaceSnoop.Wpf.Views.Controls;
+
+public partial class ScanProgressPanel : UserControl
+{
+    public ScanProgressPanel()
+    {
+        InitializeComponent();
+    }
+}
