+++ SpaceSnoop.Wpf/Views/Controls/ScanSummaryStrip.xaml
@@ -1,0 +1,142 @@
+﻿<UserControl x:Class="SpaceSnoop.Wpf.Views.Controls.ScanSummaryStrip"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:vm="clr-namespace:SpaceSnoop.Wpf.ViewModels"
+             xmlns:bootstrap="clr-namespace:SpaceSnoop.Wpf.Bootstrap"
+             xmlns:iconPacks="http://metro.mahapps.com/winfx/xaml/iconpacks"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d"
+             d:DataContext="{d:DesignInstance vm:ScanViewModel}"
+             d:DesignWidth="980">
+    <UserControl.Resources>
+        <Style x:Key="SummaryStrip"
+               TargetType="Border"
+               BasedOn="{StaticResource Card.Inset}">
+            <Setter Property="Padding"
+                    Value="0" />
+            <Setter Property="Visibility"
+                    Value="Collapsed" />
+            <Setter Property="Opacity"
+                    Value="0" />
+            <Setter Property="RenderTransform">
+                <Setter.Value>
+                    <TranslateTransform Y="8" />
+                </Setter.Value>
+            </Setter>
+            <Style.Triggers>
+                <DataTrigger Binding="{Binding HasResult}"
+                             Value="True">
+                    <Setter Property="Visibility"
+                            Value="Visible" />
+                    <DataTrigger.EnterActions>
+                        <BeginStoryboard>
+                            <Storyboard>
+                                <DoubleAnimation Storyboard.TargetProperty="Opacity"
+                                                 To="1"
+                                                 Duration="0:0:0.24" />
+                                <DoubleAnimation Storyboard.TargetProperty="RenderTransform.(TranslateTransform.Y)"
+                                                 To="0"
+                                                 Duration="0:0:0.30">
+                                    <DoubleAnimation.EasingFunction>
+                                        <CubicEase EasingMode="EaseOut" />
+                                    </DoubleAnimation.EasingFunction>
+                                </DoubleAnimation>
+                            </Storyboard>
+                        </BeginStoryboard>
+                    </DataTrigger.EnterActions>
+                </DataTrigger>
+
+                <DataTrigger Binding="{Binding IsScanning}"
+                             Value="True">
+                    <Setter Property="Visibility"
+                            Value="Collapsed" />
+                </DataTrigger>
+            </Style.Triggers>
+        </Style>
+    </UserControl.Resources>
+
+    <Border Style="{StaticResource SummaryStrip}">
+        <Grid>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="4" />
+                <ColumnDefinition Width="*" />
+                <ColumnDefinition Width="Auto" />
+            </Grid.ColumnDefinitions>
+
+            <Border Grid.Column="0"
+                    Background="{StaticResource HeatSpectrumVertical}"
+                    CornerRadius="10,0,0,10" />
+
+            <StackPanel Grid.Column="1"
+                        Orientation="Horizontal"
+                        VerticalAlignment="Center"
+                        Margin="14,7,0,7">
+                <TextBlock Text="ОБЪЁМ"
+                           Style="{StaticResource Caption}"
+                           Margin="0,0,7,0" />
+                <TextBlock Text="{Binding ResultSizeText}"
+                           Style="{StaticResource StatValue}"
+                           Foreground="{DynamicResource Accent.Primary}"
+                           Margin="0,0,20,0" />
+
+                <TextBlock Text="ФАЙЛОВ"
+                           Style="{StaticResource Caption}"
+                           Margin="0,0,7,0" />
+                <TextBlock Text="{Binding ResultFileCountText}"
+                           Style="{StaticResource StatValue}"
+                           Margin="0,0,20,0" />
+
+                <TextBlock Text="КАТАЛОГОВ"
+                           Style="{StaticResource Caption}"
+                           Margin="0,0,7,0" />
+                <TextBlock Text="{Binding ResultDirCountText}"
+                           Style="{StaticResource StatValue}"
+                           Margin="0,0,20,0" />
+
+                <TextBlock Text="ВРЕМЯ"
+                           Style="{StaticResource Caption}"
+                           Margin="0,0,7,0" />
+                <TextBlock Text="{Binding ResultElapsedText}"
+                           Style="{StaticResource StatValue}" />
+            </StackPanel>
+
+            <StackPanel Grid.Column="2"
+                        Orientation="Horizontal"
+                        VerticalAlignment="Center"
+                        Margin="20,0,14,0">
+                <iconPacks:PackIconLucide Kind="{x:Static iconPacks:PackIconLucideKind.Flame}"
+                                          Width="13"
+                                          Height="13"
+                                          VerticalAlignment="Center"
+                                          Margin="0,0,7,0"
+                                          Foreground="{DynamicResource Accent.Primary}" />
+                <TextBlock Text="ПОДСВЕТКА"
+                           Style="{StaticResource Caption}"
+                           Margin="0,0,9,0" />
+                <Border Width="60"
+                        Height="8"
+                        VerticalAlignment="Center"
+                        CornerRadius="{DynamicResource Radius.Pill}"
+                        Background="{StaticResource HeatSpectrum}" />
+                <Slider Width="116"
+                        Minimum="{x:Static bootstrap:AppDefaults.IntensityMin}"
+                        Maximum="{x:Static bootstrap:AppDefaults.IntensityMax}"
+                        Value="{Binding Intensity}"
+                        TickFrequency="1"
+                        IsSnapToTickEnabled="True"
+                        VerticalAlignment="Center"
+                        Margin="9,0,9,0"
+                        ToolTip="Чувствительность тепловой подсветки" />
+                <TextBlock Text="{Binding Intensity, StringFormat=0}"
+                           FontFamily="{DynamicResource Font.Mono}"
+                           FontSize="{DynamicResource Font.Size.M}"
+                           FontWeight="Medium"
+                           Foreground="{DynamicResource Accent.Primary}"
+                           VerticalAlignment="Center"
+                           MinWidth="18"
+                           TextAlignment="Right" />
+            </StackPanel>
+        </Grid>
+    </Border>
+</UserControl>
