+++ SpaceSnoop.Wpf/Views/Controls/ScanSummaryStrip.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace SpaceSnoop.Wpf.Views.Controls;
+
+public partial class ScanSummaryStrip : UserControl
+{
+    public ScanSummaryStrip()
+    {
+        InitializeComponent();
+    }
+}
