+++ SpaceSnoop.Wpf/Views/Controls/ScanToolbar.xaml
@@ -1,0 +1,95 @@
+﻿<UserControl x:Class="SpaceSnoop.Wpf.Views.Controls.ScanToolbar"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:vm="clr-namespace:SpaceSnoop.Wpf.ViewModels"
+             xmlns:iconPacks="http://metro.mahapps.com/winfx/xaml/iconpacks"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d"
+             d:DataContext="{d:DesignInstance vm:ScanViewModel}"
+             d:DesignWidth="980">
+    <Border Style="{StaticResource Card.Inset}"
+            Padding="14,10">
+        <Grid>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="Auto" />
+                <ColumnDefinition Width="*" />
+                <ColumnDefinition Width="Auto" />
+            </Grid.ColumnDefinitions>
+
+            <StackPanel Grid.Column="0"
+                        Orientation="Horizontal"
+                        VerticalAlignment="Center">
+                <iconPacks:PackIconLucide Kind="{x:Static iconPacks:PackIconLucideKind.HardDrive}"
+                                          Width="16"
+                                          Height="16"
+                                          VerticalAlignment="Center"
+                                          Margin="2,0,10,0"
+                                          Foreground="{DynamicResource Fg.Secondary}" />
+                <ComboBox ItemsSource="{Binding Drives}"
+                          SelectedItem="{Binding SelectedDrive}"
+                          Style="{StaticResource ComboBox.Toolbar}"
+                          MinWidth="230"
+                          Margin="0,0,8,0"
+                          ToolTip="Диск или каталог для анализа" />
+
+                <Button Style="{StaticResource Button.Secondary}"
+                        Command="{Binding BrowseCommand}"
+                        Margin="0,0,14,0"
+                        ToolTip="Выбрать каталог">
+                    <StackPanel Orientation="Horizontal">
+                        <iconPacks:PackIconLucide Kind="{x:Static iconPacks:PackIconLucideKind.FolderOpen}"
+                                                  Width="15"
+                                                  Height="15"
+                                                  VerticalAlignment="Center"
+                                                  Margin="0,0,6,0" />
+                        <TextBlock Text="Обзор" />
+                    </StackPanel>
+                </Button>
+
+                <Button Style="{StaticResource Button.Primary}"
+                        Command="{Binding StartCommand}"
+                        Margin="0,0,6,0">
+                    <StackPanel Orientation="Horizontal">
+                        <iconPacks:PackIconLucide Kind="{x:Static iconPacks:PackIconLucideKind.Play}"
+                                                  Width="15"
+                                                  Height="15"
+                                                  VerticalAlignment="Center"
+                                                  Margin="0,0,6,0" />
+                        <TextBlock Text="Сканировать" />
+                    </StackPanel>
+                </Button>
+
+                <Button Style="{StaticResource Button.Ghost}"
+                        Command="{Binding StopCommand}"
+                        ToolTip="Остановить сканирование (ESC)">
+                    <StackPanel Orientation="Horizontal">
+                        <iconPacks:PackIconLucide Kind="{x:Static iconPacks:PackIconLucideKind.Square}"
+                                                  Width="13"
+                                                  Height="13"
+                                                  VerticalAlignment="Center"
+                                                  Margin="0,0,6,0" />
+                        <TextBlock Text="Стоп" />
+                    </StackPanel>
+                </Button>
+            </StackPanel>
+
+            <StackPanel Grid.Column="2"
+                        Orientation="Horizontal"
+                        VerticalAlignment="Center">
+                <Button Style="{StaticResource Button.Destructive}"
+                        Command="{Binding DeleteMarkedCommand}"
+                        ToolTip="Переместить помеченные элементы в корзину">
+                    <StackPanel Orientation="Horizontal">
+                        <iconPacks:PackIconLucide Kind="{x:Static iconPacks:PackIconLucideKind.Trash2}"
+                                                  Width="14"
+                                                  Height="14"
+                                                  VerticalAlignment="Center"
+                                                  Margin="0,0,6,0" />
+                        <TextBlock Text="Удалить" />
+                    </StackPanel>
+                </Button>
+            </StackPanel>
+        </Grid>
+    </Border>
+</UserControl>
