+++ SpaceSnoop.Wpf/Views/Controls/ScanToolbar.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace SpaceSnoop.Wpf.Views.Controls;
+
+public partial class ScanToolbar : UserControl
+{
+    public ScanToolbar()
+    {
+        InitializeComponent();
+    }
+}
